+++ src/main/java/com/java/domain/car/Car.java
@@ -1,4 +1,4 @@
-package com.java.domain;
+package com.java.domain.car;
 
 import com.java.dto.CarDto;
